ai-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.0.1
openai==1.68.2
supabase==2.14.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20 